Move send slack notification function to config file and add notifications for approvals
